Add fields question and rating

+++ db/migrate/20190704191114_create_foundations.rb
@@ -5,7 +5,7 @@ class CreateFoundations < ActiveRecord::Migration[5.2]
       t.string :email
       t.string :city
       t.string :address
-      t.decimal :phone
+      t.integer :phone
       t.string :food
       t.time :hourEntry
       t.time :hourExit
